+++ crypto/build.info
@@ -68,6 +68,7 @@ SOURCE[../providers/fips]=$CORE_COMMON
 $UTIL_COMMON=\
         cryptlib.c mem.c mem_sec.c params.c bsearch.c ex_data.c o_str.c \
         ctype.c threads_pthread.c threads_win.c threads_none.c initthread.c \
+        punycode.c \
         context.c sparse_array.c $CPUIDASM
 $UTIL_DEFINE=$CPUIDDEF
